+++ app/modules/produk/api/proses.php
@@ -6,22 +6,20 @@ if (isset($_POST['simpan'])) {
     $kode = $_POST['kode_produk'];
     $nama = $_POST['nama_produk'];
     $kat  = $_POST['id_kategori'];
-    $beli = $_POST['harga_beli'];
-    $jual = $_POST['harga_jual'];
+    $harga = $_POST['harga'];
     $stok = $_POST['stok'];
 
-    mysqli_query($conn, "INSERT INTO produk (kode_produk, nama_produk, id_kategori, harga_beli, harga_jual, stok) VALUES ('$kode', '$nama', '$kat', '$beli', '$jual', '$stok')");
+    mysqli_query($conn, "INSERT INTO produk (kode_produk, nama_produk, id_kategori, harga, stok) VALUES ('$kode', '$nama', '$kat', '$harga', '$stok')");
 
 } elseif (isset($_POST['update'])) {
     $id   = $_POST['id_produk'];
     $kode = $_POST['kode_produk'];
     $nama = $_POST['nama_produk'];
     $kat  = $_POST['id_kategori'];
-    $beli = $_POST['harga_beli'];
-    $jual = $_POST['harga_jual'];
+    $harga = $_POST['harga'];
     $stok = $_POST['stok'];
 
-    mysqli_query($conn, "UPDATE produk SET kode_produk='$kode', nama_produk='$nama', id_kategori='$kat', harga_beli='$beli', harga_jual='$jual', stok='$stok' WHERE id_produk='$id'");
+    mysqli_query($conn, "UPDATE produk SET kode_produk='$kode', nama_produk='$nama', id_kategori='$kat', harga='$harga', stok='$stok' WHERE id_produk='$id'");
 
 } elseif (isset($_GET['act']) && $_GET['act'] == 'delete') {
     $id = $_GET['id'];